requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0

mysql-connector-python>=8.0.0
sqlalchemy>=2.0.0

pyyaml>=6.0